+++ src/assets/CollectiblesController.test.ts
@@ -58,20 +58,16 @@ describe('CollectiblesController', () => {
     collectiblesController = new CollectiblesController({
       onPreferencesStateChange: (listener) => preferences.subscribe(listener),
       onNetworkStateChange: (listener) => network.subscribe(listener),
-      getERC721AssetName: assetsContract.getERC721AssetName.bind(
-        assetsContract,
-      ),
-      getERC721AssetSymbol: assetsContract.getERC721AssetSymbol.bind(
-        assetsContract,
-      ),
+      getERC721AssetName:
+        assetsContract.getERC721AssetName.bind(assetsContract),
+      getERC721AssetSymbol:
+        assetsContract.getERC721AssetSymbol.bind(assetsContract),
       getERC721TokenURI: assetsContract.getERC721TokenURI.bind(assetsContract),
       getERC721OwnerOf: assetsContract.getERC721OwnerOf.bind(assetsContract),
-      getERC1155BalanceOf: assetsContract.getERC1155BalanceOf.bind(
-        assetsContract,
-      ),
-      getERC1155TokenURI: assetsContract.getERC1155TokenURI.bind(
-        assetsContract,
-      ),
+      getERC1155BalanceOf:
+        assetsContract.getERC1155BalanceOf.bind(assetsContract),
+      getERC1155TokenURI:
+        assetsContract.getERC1155TokenURI.bind(assetsContract),
     });
 
     preferences.update({
@@ -1294,10 +1290,11 @@ describe('CollectiblesController', () => {
           .stub(collectiblesController, 'isCollectibleOwner' as any)
           .returns(false);
 
-        const updatedCollectible = await collectiblesController.checkAndUpdateSingleCollectibleOwnershipStatus(
-          collectible,
-          true,
-        );
+        const updatedCollectible =
+          await collectiblesController.checkAndUpdateSingleCollectibleOwnershipStatus(
+            collectible,
+            true,
+          );
 
         expect(
           collectiblesController.state.allCollectibles[selectedAddress][