+++ src/assets/TokenListController.test.ts
@@ -21,9 +21,11 @@ const timestamp = Date.now();
 
 const staticTokenList: TokenListMap = {};
 for (const tokenAddress in contractMap) {
-  const { erc20, logo: filePath, ...token } = (contractMap as ContractMap)[
-    tokenAddress
-  ];
+  const {
+    erc20,
+    logo: filePath,
+    ...token
+  } = (contractMap as ContractMap)[tokenAddress];
   if (erc20) {
     staticTokenList[tokenAddress] = {
       ...token,