+++ src/keyring/KeyringController.ts
@@ -15,6 +15,10 @@ import Wallet, { thirdparty as importers } from 'ethereumjs-wallet';
 import Keyring from 'eth-keyring-controller';
 import { Mutex } from 'async-mutex';
 import {
+  MetaMaskKeyring as QRKeyring,
+  IKeyringState as IQRKeyringState,
+} from '@keystonehq/metamask-airgapped-keyring';
+import {
   BaseController,
   BaseConfig,
   BaseState,
@@ -25,14 +29,13 @@ import { PersonalMessageParams } from '../message-manager/PersonalMessageManager
 import { TypedMessageParams } from '../message-manager/TypedMessageManager';
 import { toChecksumHexAddress } from '../util';
 
-const privates = new WeakMap();
-
 /**
  * Available keyring types
  */
 export enum KeyringTypes {
   simple = 'Simple Key Pair',
   hd = 'HD Key Tree',
+  qr = 'QR Hardware Wallet Device',
 }
 
 /**
@@ -83,6 +86,7 @@ export interface KeyringMemState extends BaseState {
  */
 export interface KeyringConfig extends BaseConfig {
   encryptor?: any;
+  keyringTypes?: any[];
 }
 
 /**
@@ -130,7 +134,7 @@ export class KeyringController extends BaseController<
   /**
    * Name of this controller used during composition
    */
-  name = 'KeyringController';
+  override name = 'KeyringController';
 
   private removeIdentity: PreferencesController['removeIdentity'];
 
@@ -140,6 +144,10 @@ export class KeyringController extends BaseController<
 
   private setSelectedAddress: PreferencesController['setSelectedAddress'];
 
+  private setAccountLabel?: PreferencesController['setAccountLabel'];
+
+  #keyring: typeof Keyring;
+
   /**
    * Creates a KeyringController instance.
    *
@@ -148,6 +156,7 @@ export class KeyringController extends BaseController<
    * @param options.syncIdentities - Sync identities with the given list of addresses.
    * @param options.updateIdentities - Generate an identity for each address given that doesn't already have an identity.
    * @param options.setSelectedAddress - Set the selected address.
+   * @param options.setAccountLabel - Set a new name for account.
    * @param config - Initial options used to configure this controller.
    * @param state - Initial state to set on this controller.
    */
@@ -157,28 +166,29 @@ export class KeyringController extends BaseController<
       syncIdentities,
       updateIdentities,
       setSelectedAddress,
+      setAccountLabel,
     }: {
       removeIdentity: PreferencesController['removeIdentity'];
       syncIdentities: PreferencesController['syncIdentities'];
       updateIdentities: PreferencesController['updateIdentities'];
       setSelectedAddress: PreferencesController['setSelectedAddress'];
+      setAccountLabel?: PreferencesController['setAccountLabel'];
     },
     config?: Partial<KeyringConfig>,
     state?: Partial<KeyringState>,
   ) {
     super(config, state);
-    privates.set(this, {
-      keyring: new Keyring(Object.assign({ initState: state }, config)),
-    });
+    this.#keyring = new Keyring(Object.assign({ initState: state }, config));
 
     this.defaultState = {
-      ...privates.get(this).keyring.store.getState(),
+      ...this.#keyring.store.getState(),
       keyrings: [],
     };
     this.removeIdentity = removeIdentity;
     this.syncIdentities = syncIdentities;
     this.updateIdentities = updateIdentities;
     this.setSelectedAddress = setSelectedAddress;
+    this.setAccountLabel = setAccountLabel;
     this.initialize();
     this.fullUpdate();
   }
@@ -189,16 +199,14 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to current state when the account is added.
    */
   async addNewAccount(): Promise<KeyringMemState> {
-    const primaryKeyring = privates
-      .get(this)
-      .keyring.getKeyringsByType('HD Key Tree')[0];
+    const primaryKeyring = this.#keyring.getKeyringsByType('HD Key Tree')[0];
     /* istanbul ignore if */
     if (!primaryKeyring) {
       throw new Error('No HD keyring found');
     }
-    const oldAccounts = await privates.get(this).keyring.getAccounts();
-    await privates.get(this).keyring.addNewAccount(primaryKeyring);
-    const newAccounts = await privates.get(this).keyring.getAccounts();
+    const oldAccounts = await this.#keyring.getAccounts();
+    await this.#keyring.addNewAccount(primaryKeyring);
+    const newAccounts = await this.#keyring.getAccounts();
 
     await this.verifySeedPhrase();
 
@@ -217,14 +225,12 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to current state when the account is added.
    */
   async addNewAccountWithoutUpdate(): Promise<KeyringMemState> {
-    const primaryKeyring = privates
-      .get(this)
-      .keyring.getKeyringsByType('HD Key Tree')[0];
+    const primaryKeyring = this.#keyring.getKeyringsByType('HD Key Tree')[0];
     /* istanbul ignore if */
     if (!primaryKeyring) {
       throw new Error('No HD keyring found');
     }
-    await privates.get(this).keyring.addNewAccount(primaryKeyring);
+    await this.#keyring.addNewAccount(primaryKeyring);
     await this.verifySeedPhrase();
     return this.fullUpdate();
   }
@@ -241,10 +247,11 @@ export class KeyringController extends BaseController<
     const releaseLock = await this.mutex.acquire();
     try {
       this.updateIdentities([]);
-      const vault = await privates
-        .get(this)
-        .keyring.createNewVaultAndRestore(password, seed);
-      this.updateIdentities(await privates.get(this).keyring.getAccounts());
+      const vault = await this.#keyring.createNewVaultAndRestore(
+        password,
+        seed,
+      );
+      this.updateIdentities(await this.#keyring.getAccounts());
       this.fullUpdate();
       return vault;
     } finally {
@@ -261,10 +268,8 @@ export class KeyringController extends BaseController<
   async createNewVaultAndKeychain(password: string) {
     const releaseLock = await this.mutex.acquire();
     try {
-      const vault = await privates
-        .get(this)
-        .keyring.createNewVaultAndKeychain(password);
-      this.updateIdentities(await privates.get(this).keyring.getAccounts());
+      const vault = await this.#keyring.createNewVaultAndKeychain(password);
+      this.updateIdentities(await this.#keyring.getAccounts());
       this.fullUpdate();
       return vault;
     } finally {
@@ -278,7 +283,7 @@ export class KeyringController extends BaseController<
    * @returns Boolean returning true if the vault is unlocked.
    */
   isUnlocked(): boolean {
-    return privates.get(this).keyring.memStore.getState().isUnlocked;
+    return this.#keyring.memStore.getState().isUnlocked;
   }
 
   /**
@@ -288,8 +293,8 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to the seed phrase.
    */
   exportSeedPhrase(password: string) {
-    if (privates.get(this).keyring.password === password) {
-      return privates.get(this).keyring.keyrings[0].mnemonic;
+    if (this.#keyring.password === password) {
+      return this.#keyring.keyrings[0].mnemonic;
     }
     throw new Error('Invalid password');
   }
@@ -302,8 +307,8 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to the private key for an address.
    */
   exportAccount(password: string, address: string): Promise<string> {
-    if (privates.get(this).keyring.password === password) {
-      return privates.get(this).keyring.exportAccount(address);
+    if (this.#keyring.password === password) {
+      return this.#keyring.exportAccount(address);
     }
     throw new Error('Invalid password');
   }
@@ -314,7 +319,7 @@ export class KeyringController extends BaseController<
    * @returns A promise resolving to an array of addresses.
    */
   getAccounts(): Promise<string[]> {
-    return privates.get(this).keyring.getAccounts();
+    return this.#keyring.getAccounts();
   }
 
   /**
@@ -365,11 +370,11 @@ export class KeyringController extends BaseController<
       default:
         throw new Error(`Unexpected import strategy: '${strategy}'`);
     }
-    const newKeyring = await privates
-      .get(this)
-      .keyring.addNewKeyring(KeyringTypes.simple, [privateKey]);
+    const newKeyring = await this.#keyring.addNewKeyring(KeyringTypes.simple, [
+      privateKey,
+    ]);
     const accounts = await newKeyring.getAccounts();
-    const allAccounts = await privates.get(this).keyring.getAccounts();
+    const allAccounts = await this.#keyring.getAccounts();
     this.updateIdentities(allAccounts);
     this.setSelectedAddress(accounts[0]);
     return this.fullUpdate();
@@ -383,7 +388,7 @@ export class KeyringController extends BaseController<
    */
   async removeAccount(address: string): Promise<KeyringMemState> {
     this.removeIdentity(address);
-    await privates.get(this).keyring.removeAccount(address);
+    await this.#keyring.removeAccount(address);
     return this.fullUpdate();
   }
 
@@ -393,7 +398,7 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to current state.
    */
   setLocked(): Promise<KeyringMemState> {
-    return privates.get(this).keyring.setLocked();
+    return this.#keyring.setLocked();
   }
 
   /**
@@ -403,7 +408,7 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to a signed message string.
    */
   signMessage(messageParams: PersonalMessageParams) {
-    return privates.get(this).keyring.signMessage(messageParams);
+    return this.#keyring.signMessage(messageParams);
   }
 
   /**
@@ -413,7 +418,7 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to a signed message string.
    */
   signPersonalMessage(messageParams: PersonalMessageParams) {
-    return privates.get(this).keyring.signPersonalMessage(messageParams);
+    return this.#keyring.signPersonalMessage(messageParams);
   }
 
   /**
@@ -430,7 +435,25 @@ export class KeyringController extends BaseController<
   ) {
     try {
       const address = normalizeAddress(messageParams.from);
-      const { password } = privates.get(this).keyring;
+      const qrKeyring = await this.getOrAddQRKeyring();
+      const qrAccounts = await qrKeyring.getAccounts();
+      if (
+        qrAccounts.findIndex(
+          (qrAddress: string) =>
+            qrAddress.toLowerCase() === address.toLowerCase(),
+        ) !== -1
+      ) {
+        const messageParamsClone = { ...messageParams };
+        if (
+          version !== SignTypedDataVersion.V1 &&
+          typeof messageParamsClone.data === 'string'
+        ) {
+          messageParamsClone.data = JSON.parse(messageParamsClone.data);
+        }
+        return this.#keyring.signTypedMessage(messageParamsClone, { version });
+      }
+
+      const { password } = this.#keyring;
       const privateKey = await this.exportAccount(password, address);
       const privateKeyBuffer = toBuffer(addHexPrefix(privateKey));
       switch (version) {
@@ -463,7 +486,7 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to a signed transaction string.
    */
   signTransaction(transaction: unknown, from: string) {
-    return privates.get(this).keyring.signTransaction(transaction, from);
+    return this.#keyring.signTransaction(transaction, from);
   }
 
   /**
@@ -473,8 +496,8 @@ export class KeyringController extends BaseController<
    * @returns Promise resolving to the current state.
    */
   async submitPassword(password: string): Promise<KeyringMemState> {
-    await privates.get(this).keyring.submitPassword(password);
-    const accounts = await privates.get(this).keyring.getAccounts();
+    await this.#keyring.submitPassword(password);
+    const accounts = await this.#keyring.getAccounts();
     await this.syncIdentities(accounts);
     return this.fullUpdate();
   }
@@ -484,8 +507,8 @@ export class KeyringController extends BaseController<
    *
    * @param listener - Callback triggered when state changes.
    */
-  subscribe(listener: Listener<KeyringState>) {
-    privates.get(this).keyring.store.subscribe(listener);
+  override subscribe(listener: Listener<KeyringState>) {
+    this.#keyring.store.subscribe(listener);
   }
 
   /**
@@ -494,8 +517,8 @@ export class KeyringController extends BaseController<
    * @param listener - Callback to remove.
    * @returns True if a listener is found and unsubscribed.
    */
-  unsubscribe(listener: Listener<KeyringState>) {
-    return privates.get(this).keyring.store.unsubscribe(listener);
+  override unsubscribe(listener: Listener<KeyringState>) {
+    return this.#keyring.store.unsubscribe(listener);
   }
 
   /**
@@ -505,7 +528,7 @@ export class KeyringController extends BaseController<
    * @returns EventEmitter if listener added.
    */
   onLock(listener: () => void) {
-    return privates.get(this).keyring.on('lock', listener);
+    return this.#keyring.on('lock', listener);
   }
 
   /**
@@ -515,7 +538,7 @@ export class KeyringController extends BaseController<
    * @returns EventEmitter if listener added.
    */
   onUnlock(listener: () => void) {
-    return privates.get(this).keyring.on('unlock', listener);
+    return this.#keyring.on('unlock', listener);
   }
 
   /**
@@ -524,9 +547,7 @@ export class KeyringController extends BaseController<
    * @returns Whether the verification succeeds.
    */
   async verifySeedPhrase(): Promise<string> {
-    const primaryKeyring = privates
-      .get(this)
-      .keyring.getKeyringsByType(KeyringTypes.hd)[0];
+    const primaryKeyring = this.#keyring.getKeyringsByType(KeyringTypes.hd)[0];
     /* istanbul ignore if */
     if (!primaryKeyring) {
       throw new Error('No HD keyring found.');
@@ -539,9 +560,9 @@ export class KeyringController extends BaseController<
       throw new Error('Cannot verify an empty keyring.');
     }
 
-    const TestKeyringClass = privates
-      .get(this)
-      .keyring.getKeyringClassForType(KeyringTypes.hd);
+    const TestKeyringClass = this.#keyring.getKeyringClassForType(
+      KeyringTypes.hd,
+    );
     const testKeyring = new TestKeyringClass({
       mnemonic: seedWords,
       numberOfAccounts: accounts.length,
@@ -569,7 +590,7 @@ export class KeyringController extends BaseController<
    */
   async fullUpdate(): Promise<KeyringMemState> {
     const keyrings: Keyring[] = await Promise.all<Keyring>(
-      privates.get(this).keyring.keyrings.map(
+      this.#keyring.keyrings.map(
         async (keyring: KeyringObject, index: number): Promise<Keyring> => {
           const keyringAccounts = await keyring.getAccounts();
           const accounts = Array.isArray(keyringAccounts)
@@ -584,7 +605,125 @@ export class KeyringController extends BaseController<
       ),
     );
     this.update({ keyrings: [...keyrings] });
-    return privates.get(this).keyring.fullUpdate();
+    return this.#keyring.fullUpdate();
+  }
+
+  // QR Hardware related methods
+
+  /**
+   * Add qr hardware keyring.
+   *
+   * @returns The added keyring
+   */
+  private async addQRKeyring(): Promise<QRKeyring> {
+    const keyring = await this.#keyring.addNewKeyring(KeyringTypes.qr);
+    await this.fullUpdate();
+    return keyring;
+  }
+
+  /**
+   * Get qr hardware keyring.
+   *
+   * @returns The added keyring
+   */
+  async getOrAddQRKeyring(): Promise<QRKeyring> {
+    const keyring = this.#keyring.getKeyringsByType(KeyringTypes.qr)[0];
+    return keyring || (await this.addQRKeyring());
+  }
+
+  async restoreQRKeyring(serialized: any): Promise<void> {
+    (await this.getOrAddQRKeyring()).deserialize(serialized);
+    this.updateIdentities(await this.#keyring.getAccounts());
+    await this.fullUpdate();
+  }
+
+  async resetQRKeyringState(): Promise<void> {
+    (await this.getOrAddQRKeyring()).resetStore();
+  }
+
+  async getQRKeyringState(): Promise<IQRKeyringState> {
+    return (await this.getOrAddQRKeyring()).getMemStore();
+  }
+
+  async submitQRCryptoHDKey(cryptoHDKey: string): Promise<void> {
+    (await this.getOrAddQRKeyring()).submitCryptoHDKey(cryptoHDKey);
+  }
+
+  async submitQRCryptoAccount(cryptoAccount: string): Promise<void> {
+    (await this.getOrAddQRKeyring()).submitCryptoAccount(cryptoAccount);
+  }
+
+  async submitQRSignature(
+    requestId: string,
+    ethSignature: string,
+  ): Promise<void> {
+    (await this.getOrAddQRKeyring()).submitSignature(requestId, ethSignature);
+  }
+
+  async cancelQRSignRequest(): Promise<void> {
+    (await this.getOrAddQRKeyring()).cancelSignRequest();
+  }
+
+  async connectQRHardware(
+    page: number,
+  ): Promise<{ balance: string; address: string; index: number }[]> {
+    try {
+      const keyring = await this.getOrAddQRKeyring();
+      let accounts;
+      switch (page) {
+        case -1:
+          accounts = await keyring.getPreviousPage();
+          break;
+        case 1:
+          accounts = await keyring.getNextPage();
+          break;
+        default:
+          accounts = await keyring.getFirstPage();
+      }
+      return accounts.map((account: any) => {
+        return {
+          ...account,
+          balance: '0x0',
+        };
+      });
+    } catch (e) {
+      throw new Error(`Unspecified error when connect QR Hardware, ${e}`);
+    }
+  }
+
+  async unlockQRHardwareWalletAccount(index: number): Promise<void> {
+    const keyring = await this.getOrAddQRKeyring();
+
+    keyring.setAccountToUnlock(index);
+    const oldAccounts = await this.#keyring.getAccounts();
+    await this.#keyring.addNewAccount(keyring);
+    const newAccounts = await this.#keyring.getAccounts();
+    this.updateIdentities(newAccounts);
+    newAccounts.forEach((address: string) => {
+      if (!oldAccounts.includes(address)) {
+        if (this.setAccountLabel) {
+          this.setAccountLabel(address, `${keyring.getName()} ${index}`);
+        }
+        this.setSelectedAddress(address);
+      }
+    });
+    await this.#keyring.persistAllKeyrings();
+    await this.fullUpdate();
+  }
+
+  async getAccountKeyringType(account: string): Promise<KeyringTypes> {
+    return (await this.#keyring.getKeyringForAccount(account)).type;
+  }
+
+  async forgetQRDevice(): Promise<void> {
+    const keyring = await this.getOrAddQRKeyring();
+    keyring.forgetDevice();
+    const accounts = (await this.#keyring.getAccounts()) as string[];
+    accounts.forEach((account) => {
+      this.setSelectedAddress(account);
+    });
+    await this.#keyring.persistAllKeyrings();
+    await this.fullUpdate();
   }
 }