+++ types/eth-phishing-detect/src/detector.d.ts
@@ -1,0 +1,1 @@
+declare module 'eth-phishing-detect/src/detector';