+++ CHANGELOG.md
@@ -6,6 +6,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [28.0.0]
+### Added
+- Add GrantPermissions action to PermissionsController ([#780](https://github.com/MetaMask/controllers/pull/780))
+- Add `PermissionController.revokePermissionForAllSubjects` action ([#764](https://github.com/MetaMask/controllers/pull/764))
+
+### Changed
+- **BREAKING:** Rename NotificationController to AnnouncementController ([#697](https://github.com/MetaMask/controllers/pull/697))
+  - The `NotificationController` class is now `AnnouncementController`.
+  - The controller `notifications` state has been renamed to `announcements`.
+  - All other exported types including the word "notification" have been updated to use the word "announcement" instead.
+
+## [27.1.1]
+### Fixed
+- Move `@keystonehq/metamask-airgapped-keyring` to dependencies ([#757](https://github.com/MetaMask/controllers/pull/757))
+
+## [27.1.0] [DEPRECATED]
+### Added
+- Now the `KeyringController` supports the `QRKeyring` from `@keystonehq/metamask-airgapped-keyring`. Developers can enable the import of accounts from a QR hardware wallet. A new optional parameter, `setAccountLabel` from the `PreferencesController`, should be passed to the `KeyringController` to enable this new functionality. ([#685](https://github.com/MetaMask/controllers/pull/685))
+  - **UPDATE:** This is broken. Consumers are encouraged to upgrade to [27.1.1].
+
+### Changed
+- Bump `eth-phishing-detect` version from 1.1.14 to 1.1.16 ([#742](https://github.com/MetaMask/controllers/pull/742))
+- Bump `@metamask/contract-metadata` from 1.31.0 to 1.33.0 ([#730](https://github.com/MetaMask/controllers/pull/730))
+
+### Fixed
+- Improve error message when attempting to import an invalid private key ([#739](https://github.com/MetaMask/controllers/pull/739))
+
 ## [27.0.0]
 ### Changed
 - **BREAKING:** Further reduce load on Infura by removing non-critical data from the fallback implementation of the Gas API ([#712](https://github.com/MetaMask/controllers/pull/712))
@@ -522,7 +549,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove shapeshift controller (#209)
 
-[Unreleased]: https://github.com/MetaMask/controllers/compare/v27.0.0...HEAD
+[Unreleased]: https://github.com/MetaMask/controllers/compare/v28.0.0...HEAD
+[28.0.0]: https://github.com/MetaMask/controllers/compare/v27.1.1...v28.0.0
+[27.1.1]: https://github.com/MetaMask/controllers/compare/v27.1.0...v27.1.1
+[27.1.0]: https://github.com/MetaMask/controllers/compare/v27.0.0...v27.1.0
 [27.0.0]: https://github.com/MetaMask/controllers/compare/v26.0.0...v27.0.0
 [26.0.0]: https://github.com/MetaMask/controllers/compare/v25.1.0...v26.0.0
 [25.1.0]: https://github.com/MetaMask/controllers/compare/v25.0.0...v25.1.0